rag_chatbot/requirements.txt
flask
faiss-cpu 
sentence-transformers 
mysql-connector-python